+++ dataworkspace/dataworkspace/apps/datasets/forms.py
@@ -12,18 +12,27 @@ class FilterWidget(forms.widgets.CheckboxSelectMultiple):
         self,
         group_label,
         hint_text=None,
+        limit_initial_options=0,
+        show_more_label="Show more",
         *args,
         **kwargs  # pylint: disable=keyword-arg-before-vararg
     ):
         super().__init__(*args, **kwargs)
         self._group_label = group_label
         self._hint_text = hint_text
+        self._limit_initial_options = limit_initial_options
+        self._show_more_label = show_more_label
 
     def get_context(self, name, value, attrs):
         context = super().get_context(name, value, attrs)
         context['widget']['group_label'] = self._group_label
         context['widget']['hint_text'] = self._hint_text
+        context['widget']['limit_initial_options'] = self._limit_initial_options
+        context['widget']['show_more_label'] = self._show_more_label
         return context
+
+    class Media:
+        js = ('app-filter-show-more.js',)
 
 
 class RequestAccessForm(forms.Form):
@@ -56,11 +65,19 @@ class DatasetSearchForm(forms.Form):
             (DataSetType.VISUALISATION.value, 'View data visualisation'),
         ],
         required=False,
-        widget=FilterWidget("Purpose", hint_text="Select all that apply."),
+        widget=FilterWidget("Purpose", hint_text="What do you want to do with data?"),
     )
 
     source = forms.ModelMultipleChoiceField(
         queryset=SourceTag.objects.order_by('name').all(),
         required=False,
-        widget=FilterWidget("Source", hint_text="Select all that apply."),
+        widget=FilterWidget(
+            "Source",
+            hint_text="Source or publishing organisation",
+            limit_initial_options=10,
+            show_more_label="Show more sources",
+        ),
     )
+
+    class Media:
+        js = ('app-filter-show-more.js',)